+++ features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket/model/replyticket/TicketDetail.java
@@ -28,10 +28,10 @@ public class TicketDetail {
     private long ticketDetailId;
     @SerializedName("ticket_new_rating")
     @Expose
-    private int ticketNewRating;
+    private long ticketNewRating;
     @SerializedName("is_cs")
     @Expose
-    private int isCs;
+    private long isCs;
     @SerializedName("user_label")
     @Expose
     private String userLabel;
@@ -120,7 +120,7 @@ public class TicketDetail {
      * @param userLabelId
      *     The user_label_id
      */
-    public void setUserLabelId(int userLabelId) {
+    public void setUserLabelId(long userLabelId) {
         this.userLabelId = userLabelId;
     }
 
@@ -166,7 +166,7 @@ public class TicketDetail {
      * @param ticketDetailId
      *     The ticket_detail_id
      */
-    public void setTicketDetailId(int ticketDetailId) {
+    public void setTicketDetailId(long ticketDetailId) {
         this.ticketDetailId = ticketDetailId;
     }
 
@@ -180,7 +180,7 @@ public class TicketDetail {
      * @return
      *     The ticketNewRating
      */
-    public int getTicketNewRating() {
+    public long getTicketNewRating() {
         return ticketNewRating;
     }
 
@@ -189,7 +189,7 @@ public class TicketDetail {
      * @param ticketNewRating
      *     The ticket_new_rating
      */
-    public void setTicketNewRating(int ticketNewRating) {
+    public void setTicketNewRating(long ticketNewRating) {
         this.ticketNewRating = ticketNewRating;
     }
 
@@ -203,7 +203,7 @@ public class TicketDetail {
      * @return
      *     The isCs
      */
-    public int getIsCs() {
+    public long getIsCs() {
         return isCs;
     }
 
@@ -212,7 +212,7 @@ public class TicketDetail {
      * @param isCs
      *     The is_cs
      */
-    public void setIsCs(int isCs) {
+    public void setIsCs(long isCs) {
         this.isCs = isCs;
     }
 
@@ -304,7 +304,7 @@ public class TicketDetail {
      * @param userId
      *     The user_id
      */
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
